Release ready 2023 07 08 18 1 (#1)
* Update docker compose references

Docker For Mac has broken support for `docker-compose`. 

Since `docker compose` is a drop in replacement and it Just Works™, this PR converts `docker-compose` to `docker compose`

* Add Fig as an installation method to the README

* completion: Update completion submodule to 0.34.0

* utility: Make 'http-serve' handler faster for well known cases

In most systems, python2 or python3 command/soft-link would almost
always exist. In such cases, we don't need to invoke `python` to
detect the version. This should speed things up a bit as well.

* feat(module/ssh): add globbing for ssh-add param

fixes #2000

Signed-off-by: Shea690901 <ginny690901@hotmail.de>

---------

Signed-off-by: Shea690901 <ginny690901@hotmail.de>
Co-authored-by: Ben <benfreke@users.noreply.github.com>
Co-authored-by: Ilkin Bayramli <43158991+ibayramli2001@users.noreply.github.com>
Co-authored-by: Indrajit Raychaudhuri <irc@indrajit.com>
Co-authored-by: Shea690901 <ginny690901@hotmail.de>